+++ src/main/java/frc/robot/autos/exampleAuto.java
@@ -6,7 +6,6 @@ import frc.robot.commands.AutoArmSet;
 import frc.robot.commands.ClawCommand;
 import frc.robot.commands.LowerArmCommand;
 import frc.robot.commands.PPSwerveControllerCommand2;
-import frc.robot.subsystems.ArmToSetpoint;
 import frc.robot.subsystems.ClawSubsystem;
 import frc.robot.subsystems.LowerArmSubsystem;
 import frc.robot.subsystems.Swerve;
@@ -29,12 +28,12 @@ import edu.wpi.first.wpilibj2.command.SwerveControllerCommand;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 
 public class exampleAuto extends SequentialCommandGroup {
-    public exampleAuto(Swerve s_AutoSwerve, String path, LowerArmSubsystem sub_LowerArmSubsystem, ArmToSetpoint sub_ArmToSetpoint, int angle, ClawSubsystem sub_ClawSubsystem){
+    public exampleAuto(Swerve s_AutoSwerve, String path, LowerArmSubsystem sub_LowerArmSubsystem, int angle, ClawSubsystem sub_ClawSubsystem){
         
         PathPlannerTrajectory examplePath = PathPlanner.loadPath(path, new PathConstraints(AutoConstants.kMaxSpeedMetersPerSecond, AutoConstants.kMaxAccelerationMetersPerSecondSquared));
         // This is just an example event map. It would be better to have a constant, global event map - in your code that will be used by all path following commands.
         HashMap<String, Command> eventMap = new HashMap<>();
-        eventMap.put("setPoint", new AutoArmSet(sub_ArmToSetpoint, angle));
+
         eventMap.put("lowerArm", new LowerArmCommand(sub_LowerArmSubsystem));
         eventMap.put("claw", new ClawCommand(sub_ClawSubsystem));
