+++ src/main/java/frc/robot/commands/ArmSet.java
@@ -4,7 +4,7 @@ import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.robot.RobotContainer;
 import frc.robot.Constants.RegularConstants;
-import frc.robot.subsystems.ArmToSetpoint;
+
 import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.XboxController;
@@ -14,36 +14,24 @@ import frc.robot.RobotContainer;
 public class ArmSet extends CommandBase 
 {
     //Behold my variables
-    private final ArmToSetpoint m_UpperArmToSetpoint;
-    private final ArmToSetpoint m_LowerArmToSetpoint;
-    private boolean isFinished;
-    private double DesiredX;
-    private double DesiredY;
     
-    public ArmSet(ArmToSetpoint UpperSubsystem, ArmToSetpoint LowerSubsystem )
-    {
-        m_UpperArmToSetpoint = UpperSubsystem;
-        addRequirements(UpperSubsystem);
-        m_LowerArmToSetpoint = LowerSubsystem;
-        addRequirements(LowerSubsystem);
-    }
 
-
+boolean isFinished;
     @Override
     public void initialize() 
     {
         //puts the desired angle value on smartDashboard to be used
         SmartDashboard.putNumber("XCord", 0);
         SmartDashboard.putNumber("YCord", 0);
-        m_UpperArmToSetpoint.MoveArm(SmartDashboard.getNumber("Desired X", 0), true);
-        m_LowerArmToSetpoint.MoveArm(SmartDashboard.getNumber("Desired Y", 0), true);
+       // m_UpperArmToSetpoint.MoveArm(SmartDashboard.getNumber("Desired X", 0), true);
+       // m_LowerArmToSetpoint.MoveArm(SmartDashboard.getNumber("Desired Y", 0), true);
     }
 
 
     @Override
     public void execute() 
     {
-        DesiredX = 42;
+     /*   DesiredX = 42;
         DesiredY = 45.5;
         double L1 = RegularConstants.UpperArmLength;
         double L2 = RegularConstants.LowerArmLength;
@@ -64,7 +52,7 @@ public class ArmSet extends CommandBase
         SmartDashboard.putNumber("Upper Expected angle", UpperDesiredAngle);
         SmartDashboard.putNumber("Lower Expected Angle", LowerDesiredAngle);
         
-        Move(LowerDesiredAngle, UpperDesiredAngle);
+        */
     }
 
 
@@ -72,8 +60,7 @@ public class ArmSet extends CommandBase
     public void end(boolean interrupted) 
     {
         //stops the arm
-        m_UpperArmToSetpoint.stop();
-        m_LowerArmToSetpoint.stop();
+       
 
     }
 
@@ -81,42 +68,9 @@ public class ArmSet extends CommandBase
     @Override
     public boolean isFinished() {
       
-        return isFinished;
+        return true;
     }   
 
-    public void Move(double LowerAngle, double UpperAngle)
-    {
-        m_LowerArmToSetpoint.MoveArm(LowerAngle, false);
-        m_LowerArmToSetpoint.MoveArm(UpperAngle, true);
-
-
-        while(Math.abs(m_LowerArmToSetpoint.AngleDif)>=1 || Math.abs(m_UpperArmToSetpoint.AngleDif)>=1  )
-        {
-
-            if(Math.abs(m_LowerArmToSetpoint.AngleDif)>=1)
-            {
-                m_LowerArmToSetpoint.MoveArm(LowerAngle, false);
-            }
-
-            if(Math.abs(m_UpperArmToSetpoint.AngleDif)>=1)
-            {
-                m_UpperArmToSetpoint.MoveArm(UpperAngle, true);
-            }   
-        }
-        if(Math.abs(m_LowerArmToSetpoint.AngleDif)>=1 || Math.abs(m_UpperArmToSetpoint.AngleDif)>=1  )
-        {
-            isFinished = true;
-            m_LowerArmToSetpoint.stop();
-            return;
-            
-        }
-        else
-        {
-            isFinished = false;
-            m_LowerArmToSetpoint.stop();
-            return;
-        }
-    } 
-
+   
 }
