+++ src/main/java/frc/robot/commands/AutoArmSet.java
@@ -5,7 +5,7 @@ import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.robot.Constants;
 import frc.robot.RobotContainer;
 import frc.robot.Constants.RegularConstants;
-import frc.robot.subsystems.ArmToSetpoint;
+//import frc.robot.subsystems.ArmToSetpoint;
 import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.Timer;
@@ -16,7 +16,7 @@ import frc.robot.RobotContainer;
 public class AutoArmSet extends CommandBase 
 {
     //Behold my variables
-    private final ArmToSetpoint m_ArmToSetpoint;
+   // private final ArmToSetpoint m_ArmToSetpoint;
     private boolean isFinished;
     public double LastDesiredAngle;
     public double DesiredAngle = 0;
@@ -25,10 +25,9 @@ public class AutoArmSet extends CommandBase
     DigitalInput limitSwitch = new DigitalInput(5);
     private int angle;
 
-    public AutoArmSet(ArmToSetpoint subsystem, int angle)
+    public AutoArmSet(int angle)
     {
-        m_ArmToSetpoint = subsystem;
-        addRequirements(subsystem);
+       
         this.angle = angle;
     }
 
@@ -38,7 +37,7 @@ public class AutoArmSet extends CommandBase
     public void initialize() 
     {
         //puts the desired angle value on smartDashboard to be used
-        m_ArmToSetpoint.MoveArm(SmartDashboard.getNumber("Desired Angle", 0), true);
+      //  m_ArmToSetpoint.MoveArm(SmartDashboard.getNumber("Desired Angle", 0), true);
     }
 
 
@@ -56,8 +55,7 @@ public class AutoArmSet extends CommandBase
     @Override
     public void end(boolean interrupted) 
     {
-        //stops the arm
-        m_ArmToSetpoint.stop();
+ 
     }
 
     
@@ -73,42 +71,9 @@ public class AutoArmSet extends CommandBase
 
     public void setAngle()
     {
-        Move(angle, true);
-        LastDesiredAngle = DesiredAngle;
+      
     }
     
    
 
-    public void Move(double angle, boolean PID)
-    {
-        m_ArmToSetpoint.MoveArm(angle, true);
-        while(Math.abs(m_ArmToSetpoint.AngleDif)>=1)
-        {
-            if(limitSwitch.get()==true)
-            {
-                DesiredAngle = 0;
-            
-            }
-            else
-            {
-                m_ArmToSetpoint.MoveArm(angle, PID);
-            }
-            
-        }
-        if(Math.abs(m_ArmToSetpoint.AngleDif)>=1)
-        {
-            isFinished = true;
-            m_ArmToSetpoint.stop();
-            return;
-            
-        }
-        else
-        {
-            isFinished = false;
-            m_ArmToSetpoint.stop();
-            return;
-        }
-    } 
 }
-
-
