+++ src/main/java/frc/robot/commands/MoveArmManualCommand.java
@@ -8,7 +8,9 @@ import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
-import frc.robot.subsystems.ArmToSetpoint;
+import frc.robot.subsystems.EncoderCheck;
+import frc.robot.subsystems.LowerArmSubsystem;
+import frc.robot.subsystems.UpperArmManual;
 
 public class MoveArmManualCommand extends CommandBase {
   /** Creates a new MoveArmManualCommand. */
@@ -18,41 +20,84 @@ public class MoveArmManualCommand extends CommandBase {
   public final JoystickButton UpperArmDecreaseButton = new JoystickButton(arm, XboxController.Button.kRightBumper.value); //works
   private final JoystickButton lowerArmIncreaseButton = new JoystickButton(arm, XboxController.Button.kBack.value); //works
   private final JoystickButton lowerArmDecreaseButton = new JoystickButton(arm, XboxController.Button.kStart.value); //works
-  private final ArmToSetpoint m_lowerArmToSetpoint;
-  private final ArmToSetpoint m_upperArmToSetpoint;
-  public MoveArmManualCommand(ArmToSetpoint UpperArmToSetpoint, ArmToSetpoint LowerArmToSetpoint) {
+  private final UpperArmManual m_UpperArmManual;
+  private final LowerArmSubsystem m_LowerArmSubsystem;
+  private final EncoderCheck m_EncoderCheck;
+  public MoveArmManualCommand(UpperArmManual upperArmManual, LowerArmSubsystem lowerArmSubsystem,EncoderCheck encoderCheck) {
     // Use addRequirements() here to declare subsystem dependencies.
-    m_lowerArmToSetpoint = LowerArmToSetpoint;
-    m_upperArmToSetpoint = UpperArmToSetpoint;
-    addRequirements(UpperArmToSetpoint);
-    addRequirements(LowerArmToSetpoint);
+ 
+    m_UpperArmManual = upperArmManual;
+    m_LowerArmSubsystem = lowerArmSubsystem;
+    m_EncoderCheck = encoderCheck;
+    addRequirements(upperArmManual);
+    addRequirements(upperArmManual);
+    addRequirements(lowerArmSubsystem);
+
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
+  public void initialize() 
+  {
+    m_LowerArmSubsystem.m_initializeSolenoid();
+    m_LowerArmSubsystem.m_enableCompressor();
+  }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() 
   {
-    int lowerTemp_EncoderCheck = m_lowerArmToSetpoint.EncoderCheck();
-    int upperTemp_EncoderCheck = m_upperArmToSetpoint.EncoderCheck();
+    int upperTemp_EncoderCheck = m_EncoderCheck.CheckEncoders(true);
+    int lowerTemp_EncoderCheck = m_EncoderCheck.CheckEncoders(false);
+
+    //Upper Arm Move Down
     if(UpperArmDecreaseButton.getAsBoolean())
     {
-      m_upperArmToSetpoint.moveDown(.3, upperTemp_EncoderCheck);
+      if(upperTemp_EncoderCheck==3)
+      {
+        return;
+      }
+      else if(upperTemp_EncoderCheck == 2)
+      {
+        m_UpperArmManual.moveDown(.1);
+        return;
+      }
+      m_UpperArmManual.moveDown(.3);
     }
+
+    //Upper Arm Move Up
     else if(UpperArmIncreaseButton.getAsBoolean())
     {
-      m_upperArmToSetpoint.moveUp(.3,upperTemp_EncoderCheck);
+      if(upperTemp_EncoderCheck == 1)
+      {
+        m_UpperArmManual.moveDown(.2);
+        return;
+      }
+      else if(upperTemp_EncoderCheck == 2)
+      {
+        m_UpperArmManual.moveUp(.1);
+        return;
+      }
+      m_UpperArmManual.moveUp(.3);
     }
+
+    //Lower Arm Move Down
     else if(lowerArmDecreaseButton.getAsBoolean())
     {
-      m_lowerArmToSetpoint.moveDown(.3,lowerTemp_EncoderCheck);
+      m_LowerArmSubsystem.m_contract();
     }
+
+    //Lower Arm Move Up
     else if(lowerArmIncreaseButton.getAsBoolean())
     {
-      m_lowerArmToSetpoint.moveUp(.3, lowerTemp_EncoderCheck);
+      if(lowerTemp_EncoderCheck ==11)
+      {
+        return;
+      }
+      else
+      {
+        m_LowerArmSubsystem.m_extend();
+      }
     }
   }
